refactor(backend-automation): split modules
