+++ src/backend/automation.rs
@@ -54,6 +54,12 @@ use crate::{
     workload_engine::{WorkloadEngineManager, WorkloadEngineSnapshot, WorkloadEngineUpdate},
 };
 
+mod requirements;
+mod status;
+
+use requirements::*;
+use status::*;
+
 const ACTIVE_PLAN_REFRESH_INTERVAL: Duration = Duration::from_secs(10);
 const CPU_USAGE_REFRESH_INTERVAL: Duration = Duration::from_secs(1);
 const ECO_QOS_REFRESH_INTERVAL: Duration = Duration::from_secs(1);
@@ -975,736 +981,6 @@ fn run_background_automation(shared: Arc<SharedAutomationState>) {
 
 fn min_worker_wait(current: Option<Duration>, candidate: Duration) -> Duration {
     current.map_or(candidate, |current| current.min(candidate))
-}
-
-struct AutomationSnapshot {
-    settings: Arc<Settings>,
-    change_generation: u64,
-    app_suspension_freeze_requests: Vec<String>,
-    memory_trim_now_requested: bool,
-    action_log_clear_requested: bool,
-    wake_events: AutomationWakeEvents,
-    windows_event_watcher_active: bool,
-}
-
-fn automation_snapshot(shared: &SharedAutomationState) -> Option<AutomationSnapshot> {
-    shared.state.lock().ok().and_then(|mut state| {
-        (!state.stop_requested).then(|| AutomationSnapshot {
-            settings: state.settings.clone(),
-            change_generation: state.change_generation,
-            app_suspension_freeze_requests: std::mem::take(
-                &mut state.app_suspension_freeze_requests,
-            ),
-            memory_trim_now_requested: std::mem::take(&mut state.memory_trim_now_requested),
-            action_log_clear_requested: std::mem::take(&mut state.action_log_clear_requested),
-            wake_events: std::mem::take(&mut state.pending_events),
-            windows_event_watcher_active: state.windows_event_watcher_active,
-        })
-    })
-}
-
-fn set_windows_event_watcher_active(shared: &SharedAutomationState, active: bool) {
-    if let Ok(mut state) = shared.state.lock() {
-        if state.windows_event_watcher_active == active {
-            return;
-        }
-
-        state.windows_event_watcher_active = active;
-        state.change_generation = state.change_generation.wrapping_add(1);
-        shared.changed.notify_one();
-    }
-}
-
-fn notify_windows_event(shared: &SharedAutomationState, event: WindowsAutomationEvent) {
-    if let Ok(mut state) = shared.state.lock() {
-        if state.stop_requested || !windows_event_wake_required(&state.settings, event) {
-            return;
-        }
-
-        if event == WindowsAutomationEvent::AppearanceChanged {
-            state.appearance_change_generation = state.appearance_change_generation.wrapping_add(1);
-            bump_status_generation(shared, &mut state);
-        }
-        state.pending_events.insert_windows_event(event);
-        state.change_generation = state.change_generation.wrapping_add(1);
-        shared.changed.notify_one();
-    }
-}
-
-fn notify_input_event(shared: &SharedAutomationState, events: InputHookEvents) {
-    if let Ok(mut state) = shared.state.lock() {
-        if state.stop_requested || !input_hook_should_check(&state.settings, events) {
-            return;
-        }
-
-        if input_hook_should_check_activity(&state.settings, events) {
-            state.pending_events.input_activity = true;
-        }
-        if input_hook_should_check_app_switch(&state.settings, events) {
-            state.pending_events.app_switch = true;
-        }
-        if input_hook_should_check_app_switch_mouse_click(&state.settings, events) {
-            state.pending_events.app_switch_mouse_click = true;
-        }
-        state.change_generation = state.change_generation.wrapping_add(1);
-        shared.changed.notify_one();
-    }
-}
-
-fn update_background_efficiency_status(
-    shared: &SharedAutomationState,
-    status: BackgroundEfficiencySnapshot,
-) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.background_efficiency,
-        |state| &mut state.background_efficiency_status,
-    );
-}
-
-fn update_app_suspension_status(shared: &SharedAutomationState, status: AppSuspensionSnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.app_suspension,
-        |state| &mut state.app_suspension_status,
-    );
-}
-
-fn update_core_steering_status(shared: &SharedAutomationState, status: CoreSteeringSnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.core_steering,
-        |state| &mut state.core_steering_status,
-    );
-}
-
-fn update_background_cpu_restriction_status(
-    shared: &SharedAutomationState,
-    status: CoreSteeringSnapshot,
-) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.background_cpu_restriction,
-        |state| &mut state.background_cpu_restriction_status,
-    );
-}
-
-fn append_unique_process_names(target: &mut Vec<String>, names: &[String]) -> bool {
-    let old_len = target.len();
-    for name in names {
-        let name = process_name_key(name);
-        if !name.is_empty()
-            && !target
-                .iter()
-                .any(|existing| existing.eq_ignore_ascii_case(&name))
-        {
-            target.push(name);
-        }
-    }
-    target.len() != old_len
-}
-
-fn update_core_limiter_status(shared: &SharedAutomationState, status: CoreLimiterSnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.core_limiter,
-        |state| &mut state.core_limiter_status,
-    );
-}
-
-fn update_by_running_app_status(shared: &SharedAutomationState, status: ByRunningAppSnapshot) {
-    update_status(shared, status, |state| &mut state.by_running_app_status);
-}
-
-fn update_workload_engine_status(shared: &SharedAutomationState, status: WorkloadEngineSnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.workload_engine,
-        |state| &mut state.workload_engine_status,
-    );
-}
-
-fn update_io_priority_status(shared: &SharedAutomationState, status: IoPrioritySnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.io_priority,
-        |state| &mut state.io_priority_status,
-    );
-}
-
-fn update_process_priority_status(shared: &SharedAutomationState, status: ProcessPrioritySnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.process_priority,
-        |state| &mut state.process_priority_status,
-    );
-}
-
-fn update_thread_priority_status(shared: &SharedAutomationState, status: ThreadPrioritySnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.thread_priority,
-        |state| &mut state.thread_priority_status,
-    );
-}
-
-fn update_dynamic_priority_boost_status(
-    shared: &SharedAutomationState,
-    status: DynamicPriorityBoostSnapshot,
-) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.dynamic_priority_boost,
-        |state| &mut state.dynamic_priority_boost_status,
-    );
-}
-
-fn update_gpu_priority_status(shared: &SharedAutomationState, status: GpuPrioritySnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.gpu_priority,
-        |state| &mut state.gpu_priority_status,
-    );
-}
-
-fn update_memory_priority_status(shared: &SharedAutomationState, status: MemoryPrioritySnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.memory_priority,
-        |state| &mut state.memory_priority_status,
-    );
-}
-
-fn update_memory_trim_status(shared: &SharedAutomationState, status: MemoryTrimSnapshot) {
-    update_status_with_auto_exclusions(
-        shared,
-        status.clone(),
-        &status.auto_excluded_processes,
-        |pending| &mut pending.memory_trim,
-        |state| &mut state.memory_trim_status,
-    );
-}
-
-fn update_timer_resolution_status(shared: &SharedAutomationState, status: TimerResolutionSnapshot) {
-    update_status(shared, status, |state| &mut state.timer_resolution_status);
-}
-
-fn update_status<T: PartialEq>(
-    shared: &SharedAutomationState,
-    status: T,
-    field: impl for<'a> FnOnce(&'a mut AutomationWorkerState) -> &'a mut T,
-) {
-    if let Ok(mut state) = shared.state.lock() {
-        if set_status(field(&mut state), status) {
-            bump_status_generation(shared, &mut state);
-        }
-    }
-}
-
-fn update_status_with_auto_exclusions<T: PartialEq>(
-    shared: &SharedAutomationState,
-    status: T,
-    auto_excluded_processes: &[String],
-    pending_field: impl for<'a> FnOnce(&'a mut PendingAutoExclusions) -> &'a mut Vec<String>,
-    status_field: impl for<'a> FnOnce(&'a mut AutomationWorkerState) -> &'a mut T,
-) {
-    if let Ok(mut state) = shared.state.lock() {
-        if append_unique_process_names(
-            pending_field(&mut state.pending_auto_exclusions),
-            auto_excluded_processes,
-        ) {
-            shared
-                .pending_auto_exclusions_generation
-                .fetch_add(1, Ordering::Release);
-        }
-        if set_status(status_field(&mut state), status) {
-            bump_status_generation(shared, &mut state);
-        }
-    }
-}
-
-fn update_action_log_entries(shared: &SharedAutomationState, entries: Vec<ActionLogEntry>) {
-    if let Ok(mut state) = shared.state.lock() {
-        let entries = Arc::new(entries);
-        if state.action_log_entries != entries {
-            state.action_log_entries = entries;
-            bump_status_generation(shared, &mut state);
-        }
-    }
-}
-
-fn set_status<T: PartialEq>(current: &mut T, next: T) -> bool {
-    if *current != next {
-        *current = next;
-        true
-    } else {
-        false
-    }
-}
-
-fn bump_status_generation(shared: &SharedAutomationState, state: &mut AutomationWorkerState) {
-    state.status_generation = state.status_generation.wrapping_add(1);
-    shared
-        .status_generation
-        .store(state.status_generation, Ordering::Release);
-}
-
-fn automation_refresh_interval(
-    hidden_to_tray: bool,
-    adaptive_engine_enabled: bool,
-    hidden_interval: Duration,
-) -> Duration {
-    // ponytail: one global saver cadence; add per-feature intervals only if a real workflow needs it.
-    if adaptive_engine_enabled {
-        hidden_interval.max(ADAPTIVE_ENGINE_AUTOMATION_REFRESH_INTERVAL)
-    } else if hidden_to_tray {
-        hidden_interval.max(HIDDEN_AUTOMATION_REFRESH_INTERVAL)
-    } else {
-        VISIBLE_AUTOMATION_REFRESH_INTERVAL
-    }
-}
-
-fn workload_refresh_interval(
-    settings: &Settings,
-    hidden_to_tray: bool,
-    adaptive_engine_enabled: bool,
-) -> Duration {
-    if adaptive_power_plan_required(settings) {
-        WORKLOAD_ENGINE_FAST_REFRESH_INTERVAL
-    } else {
-        automation_refresh_interval(
-            hidden_to_tray,
-            adaptive_engine_enabled,
-            WORKLOAD_ENGINE_REFRESH_INTERVAL,
-        )
-    }
-}
-
-fn memory_trim_refresh_interval(settings: &Settings) -> Duration {
-    Duration::from_secs(
-        settings
-            .memory_trim
-            .check_interval_minutes
-            .max(1)
-            .saturating_mul(60),
-    )
-}
-
-fn workload_engine_fast_refresh_deadline(settings: &Settings, now: Instant) -> Option<Instant> {
-    feature_refresh_required(settings, workload_engine_required(settings))
-        .then_some(now + WORKLOAD_ENGINE_FAST_REFRESH_WINDOW)
-}
-
-fn workload_engine_fast_refresh_active(
-    settings: &Settings,
-    fast_until: Option<Instant>,
-    now: Instant,
-) -> bool {
-    feature_refresh_required(settings, workload_engine_required(settings))
-        && fast_until.is_some_and(|until| now < until)
-}
-
-fn feature_refresh_required(settings: &Settings, feature_enabled: bool) -> bool {
-    settings.general.enabled && feature_enabled
-}
-
-fn workload_engine_required(settings: &Settings) -> bool {
-    let workload = &settings.workload_engine;
-    workload.enabled
-        && (workload.lower_background_apps
-            || workload.workload_engine_background_efficiency_enabled
-            || workload.workload_engine_enabled
-            || workload.boost_foreground_app)
-}
-
-fn io_priority_required(settings: &Settings) -> bool {
-    settings.io_priority.enabled
-        || (settings.workload_engine.enabled
-            && (settings
-                .workload_engine
-                .lower_background_io_priority_enabled
-                || settings.workload_engine.workload_engine_io_priority.enabled))
-}
-
-fn workload_engine_priority_assist_required(settings: &Settings) -> bool {
-    settings.workload_engine.enabled && settings.workload_engine.workload_engine_enabled
-}
-
-fn thread_priority_required(settings: &Settings) -> bool {
-    settings.thread_priority.enabled
-        || (workload_engine_priority_assist_required(settings)
-            && settings
-                .workload_engine
-                .workload_engine_thread_priority
-                .enabled)
-}
-
-fn dynamic_priority_boost_required(settings: &Settings) -> bool {
-    settings.dynamic_priority_boost.enabled
-        || (workload_engine_priority_assist_required(settings)
-            && settings
-                .workload_engine
-                .workload_engine_dynamic_priority_boost
-                .enabled)
-}
-
-fn gpu_priority_required(settings: &Settings) -> bool {
-    settings.gpu_priority.enabled
-        || (workload_engine_priority_assist_required(settings)
-            && settings
-                .workload_engine
-                .workload_engine_gpu_priority
-                .enabled)
-}
-
-fn memory_priority_required(settings: &Settings) -> bool {
-    settings.memory_priority.enabled
-}
-
-fn timer_resolution_required(settings: &Settings) -> bool {
-    settings.timer_resolution.enabled
-}
-
-fn effective_io_priority_settings(
-    settings: &Settings,
-    _launch_boost_active: bool,
-    workload_engine_active: bool,
-) -> crate::config::IoPrioritySettings {
-    let mut io_priority = settings.io_priority.clone();
-    if workload_engine_active {
-        let auto_io_priority = workload_engine_io_priority_settings(settings);
-        if auto_io_priority.enabled {
-            io_priority = auto_io_priority;
-            io_priority
-                .exclusions
-                .extend(settings.workload_engine.workload_engine_exclusions.clone());
-        }
-    }
-    io_priority
-}
-
-fn workload_engine_io_priority_settings(settings: &Settings) -> crate::config::IoPrioritySettings {
-    let mut io_priority = settings.workload_engine.workload_engine_io_priority.clone();
-    if !io_priority.enabled
-        && settings
-            .workload_engine
-            .lower_background_io_priority_enabled
-    {
-        io_priority.enabled = true;
-        io_priority.foreground_priority = ProcessIoPriority::Normal.into();
-        io_priority.background_priority =
-            settings.workload_engine.lower_background_io_priority.into();
-    }
-    io_priority.foreground_detection_enabled = true;
-    io_priority.preserve_foreground_priority = true;
-    io_priority.preserve_background_priority = true;
-    io_priority
-}
-
-fn effective_thread_priority_settings(
-    settings: &Settings,
-    workload_engine_active: bool,
-) -> crate::config::ThreadPrioritySettings {
-    let mut thread_priority = settings.thread_priority.clone();
-    if workload_engine_active
-        && workload_engine_priority_assist_required(settings)
-        && settings
-            .workload_engine
-            .workload_engine_thread_priority
-            .enabled
-    {
-        thread_priority = settings
-            .workload_engine
-            .workload_engine_thread_priority
-            .clone();
-        thread_priority.foreground_detection_enabled = true;
-        thread_priority.preserve_foreground_priority = true;
-        thread_priority.preserve_background_priority = true;
-        thread_priority
-            .exclusions
-            .extend(settings.workload_engine.workload_engine_exclusions.clone());
-    }
-    thread_priority
-}
-
-fn effective_dynamic_priority_boost_settings(
-    settings: &Settings,
-    workload_engine_active: bool,
-) -> crate::config::DynamicPriorityBoostSettings {
-    let mut dynamic_priority_boost = settings.dynamic_priority_boost.clone();
-    if workload_engine_active
-        && workload_engine_priority_assist_required(settings)
-        && settings
-            .workload_engine
-            .workload_engine_dynamic_priority_boost
-            .enabled
-    {
-        dynamic_priority_boost = settings
-            .workload_engine
-            .workload_engine_dynamic_priority_boost
-            .clone();
-        dynamic_priority_boost.foreground_detection_enabled = true;
-        dynamic_priority_boost
-            .exclusions
-            .extend(settings.workload_engine.workload_engine_exclusions.clone());
-    }
-    dynamic_priority_boost
-}
-
-fn effective_gpu_priority_settings(
-    settings: &Settings,
-    workload_engine_active: bool,
-) -> crate::config::GpuPrioritySettings {
-    let mut gpu_priority = settings.gpu_priority.clone();
-    if workload_engine_active
-        && workload_engine_priority_assist_required(settings)
-        && settings
-            .workload_engine
-            .workload_engine_gpu_priority
-            .enabled
-    {
-        gpu_priority = settings
-            .workload_engine
-            .workload_engine_gpu_priority
-            .clone();
-        gpu_priority.foreground_detection_enabled = true;
-        gpu_priority.preserve_foreground_priority = true;
-        gpu_priority.preserve_background_priority = true;
-        gpu_priority
-            .exclusions
-            .extend(settings.workload_engine.workload_engine_exclusions.clone());
-    }
-    gpu_priority
-}
-
-fn effective_memory_priority_settings(
-    settings: &Settings,
-) -> crate::config::MemoryPrioritySettings {
-    settings.memory_priority.clone()
-}
-
-fn process_appearance_scan_required(settings: &Settings) -> bool {
-    settings.general.enabled
-        && (settings.background_efficiency.enabled
-            || settings.core_steering.enabled
-            || settings.background_cpu_restriction.enabled
-            || settings.core_limiter.enabled
-            || settings.by_running_app.enabled
-            || settings.workload_engine.enabled
-            || settings.process_priority.enabled
-            || thread_priority_required(settings)
-            || dynamic_priority_boost_required(settings)
-            || io_priority_required(settings)
-            || gpu_priority_required(settings)
-            || memory_priority_required(settings)
-            || settings.memory_trim.enabled)
-}
-
-fn power_plan_checks_required(settings: &Settings) -> bool {
-    settings.general.enabled
-        && (activity_power_plan_required(settings)
-            || by_foreground_required(settings)
-            || by_time_rules_required(settings)
-            || by_cpu_load_rules_required(settings)
-            || by_running_app_required(settings))
-}
-
-fn automation_worker_required(settings: &Settings) -> bool {
-    settings.general.enabled
-        && (power_plan_checks_required(settings)
-            || adaptive_power_plan_required(settings)
-            || settings.background_efficiency.enabled
-            || settings.app_suspension.enabled
-            || settings.core_steering.enabled
-            || settings.background_cpu_restriction.enabled
-            || settings.core_limiter.enabled
-            || settings.by_running_app.enabled
-            || settings.workload_engine.enabled
-            || settings.process_priority.enabled
-            || thread_priority_required(settings)
-            || dynamic_priority_boost_required(settings)
-            || io_priority_required(settings)
-            || gpu_priority_required(settings)
-            || memory_priority_required(settings)
-            || settings.memory_trim.enabled
-            || timer_resolution_required(settings))
-}
-
-fn windows_event_watcher_required(settings: &Settings) -> bool {
-    automation_windows_event_watcher_required(settings)
-        || (!settings.adaptive_engine.enabled && appearance_events_required(settings))
-}
-
-fn automation_windows_event_watcher_required(settings: &Settings) -> bool {
-    settings.general.enabled
-        && (power_plan_checks_required(settings) || event_driven_process_work_required(settings))
-}
-
-fn event_driven_process_work_required(settings: &Settings) -> bool {
-    !settings.adaptive_engine.enabled
-        && (settings.app_suspension.enabled || process_appearance_scan_required(settings))
-}
-
-fn windows_event_wake_required(settings: &Settings, event: WindowsAutomationEvent) -> bool {
-    if event == WindowsAutomationEvent::AppearanceChanged {
-        return !settings.adaptive_engine.enabled && appearance_events_required(settings);
-    }
-
-    if settings.general.enabled {
-        match event {
-            WindowsAutomationEvent::ForegroundChanged => {
-                power_plan_checks_required(settings) || event_driven_process_work_required(settings)
-            }
-            WindowsAutomationEvent::WindowCreated => event_driven_process_work_required(settings),
-            WindowsAutomationEvent::PowerChanged => power_plan_checks_required(settings),
-            WindowsAutomationEvent::SessionChanged => windows_event_watcher_required(settings),
-            WindowsAutomationEvent::AppearanceChanged => false,
-        }
-    } else {
-        false
-    }
-}
-
-fn appearance_events_required(settings: &Settings) -> bool {
-    settings.general.theme_mode == AppThemeMode::System
-        || settings.general.accent.source == AccentColorSource::Windows
-        || settings.general.animation_mode == AnimationMode::System
-}
-
-fn activity_power_plan_required(settings: &Settings) -> bool {
-    settings.by_activity.enabled
-        && (has_idle_plan(&settings.by_activity.power_plans)
-            || (settings.by_activity.switch_to_performance_on_resume
-                && settings.by_activity.input_detection.any_enabled()
-                && has_active_plan(&settings.by_activity.power_plans)))
-}
-
-fn controller_activity_poll_required(settings: &Settings) -> bool {
-    settings.general.enabled
-        && settings.by_activity.enabled
-        && settings.by_activity.input_detection.controller
-        && (has_idle_plan(&settings.by_activity.power_plans)
-            || has_active_plan(&settings.by_activity.power_plans))
-}
-
-fn by_foreground_required(settings: &Settings) -> bool {
-    settings.by_foreground.enabled
-        && (settings
-            .by_foreground
-            .rules
-            .iter()
-            .any(|rule| rule.enabled && rule.power_plan_guid.is_some()))
-}
-
-fn foreground_lookup_required(settings: &Settings) -> bool {
-    settings.by_foreground.enabled && !settings.by_foreground.rules.is_empty()
-}
-
-fn by_time_rules_required(settings: &Settings) -> bool {
-    settings.by_time.enabled
-        && settings
-            .by_time
-            .rules
-            .iter()
-            .any(|rule| rule.enabled && rule.power_plan_guid.is_some())
-}
-
-fn by_cpu_load_rules_required(settings: &Settings) -> bool {
-    settings.by_cpu_load.enabled
-        && settings.by_cpu_load.rules.iter().any(|rule| {
-            rule.enabled
-                && (rule.power_plan_guid.is_some()
-                    || (rule.else_enabled && rule.else_power_plan_guid.is_some()))
-        })
-}
-
-fn by_running_app_required(settings: &Settings) -> bool {
-    settings.by_running_app.enabled
-        && settings
-            .by_running_app
-            .rules
-            .iter()
-            .any(|rule| rule.enabled && rule.power_plan_guid.is_some())
-}
-
-fn has_idle_plan(power_plans: &PowerPlanSettings) -> bool {
-    power_plans.power_save_guid.is_some()
-}
-
-fn has_active_plan(power_plans: &PowerPlanSettings) -> bool {
-    power_plans.performance_guid.is_some()
-}
-
-fn configured_check_interval(settings: &Settings) -> Duration {
-    Duration::from_millis(settings.general.check_interval_ms.max(250))
-}
-
-fn hidden_power_plan_check_delay(
-    settings: &Settings,
-    windows_event_watcher_active: bool,
-) -> Option<Duration> {
-    if !windows_event_watcher_active {
-        return Some(configured_check_interval(settings));
-    }
-
-    let mut delay = None;
-    if by_cpu_load_rules_required(settings) {
-        delay = Some(min_worker_wait(delay, CPU_USAGE_REFRESH_INTERVAL));
-    }
-    if by_time_rules_required(settings) {
-        let schedule_delay = ByTimeScheduler
-            .next_change_delay(&settings.by_time)
-            .map(|delay| delay.min(SCHEDULE_RULE_MAX_SLEEP))
-            .unwrap_or_else(|| configured_check_interval(settings));
-        delay = Some(min_worker_wait(delay, schedule_delay));
-    }
-    if by_running_app_required(settings) {
-        delay = Some(min_worker_wait(delay, PERFORMANCE_MODE_REFRESH_INTERVAL));
-    }
-    if let Some(activity_delay) = activity_idle_check_delay(settings) {
-        delay = Some(min_worker_wait(delay, activity_delay));
-    }
-    delay
-}
-
-fn activity_idle_check_delay(settings: &Settings) -> Option<Duration> {
-    if !settings.general.enabled
-        || !settings.by_activity.enabled
-        || !has_idle_plan(&settings.by_activity.power_plans)
-    {
-        return None;
-    }
-
-    let timeout = Duration::from_secs(settings.by_activity.idle_timeout_seconds);
-    match input_tracker::last_input_elapsed() {
-        Some(idle_for) if idle_for < timeout => Some(timeout - idle_for),
-        Some(_) => None,
-        None => Some(configured_check_interval(settings)),
-    }
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
